fix services and tests
