+++ backend/python/app/models/__init__.py
@@ -12,13 +12,13 @@ def init_app(app):
         caregiver,
         child,
         child_behavior,
-        children_child_behaviors,
+        # children_child_behaviors,
         daytime_contact,
         familial_concern,
         goal,
         intake,
-        intakes_concerns,
-        intakes_goals,
+        # intakes_concerns,
+        # intakes_goals,
         other_permitted_individual,
         provider,
         transportation_method,
